Refactor transaction.py phase 5
