+++ counterparty-core/counterpartycore/test/util_test.py
@@ -43,6 +43,7 @@ from counterpartycore.lib import (  # noqa: E402
     transaction,
     util,
 )
+from counterpartycore.lib.api.util import to_json  # noqa: E402
 from counterpartycore.lib.messages import fairminter, utxo  # noqa
 from counterpartycore.test.fixtures.params import DEFAULT_PARAMS as DP  # noqa: E402
 from counterpartycore.test.fixtures.scenarios import (  # noqa: E402
@@ -911,9 +912,7 @@ def check_outputs(
                     )
                 else:
                     msg = f"expected outputs don't match test_outputs: expected_outputs={outputs} test_outputs={test_outputs}"
-                import json
-
-                print(json.dumps(test_outputs, indent=4))
+                print(to_json(test_outputs))
                 raise Exception(msg)  # noqa: B904
         if records is not None:
             for record in records:
