+++ counterparty-core/counterpartycore/test/utxolocks_test.py
@@ -7,6 +7,7 @@ import bitcoin
 
 from counterpartycore.lib import transaction
 from counterpartycore.lib.messages import send
+from counterpartycore.lib.transaction_helper import transaction_inputs
 from counterpartycore.test import (
     conftest,  # noqa: F401
 )
@@ -27,7 +28,7 @@ def construct_tx(db, source, destination, disable_utxo_locks=False, inputs_set=N
 
 
 def test_utxolocks(server_db):
-    transaction.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
+    transaction_inputs.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
 
     """it shouldn't use the same UTXO"""
     tx1hex = construct_tx(
@@ -50,7 +51,7 @@ def test_utxolocks(server_db):
 
 
 def test_utxolocks_custom_input(server_db):
-    transaction.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
+    transaction_inputs.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
 
     """it should use the same UTXO"""
     inputs_set = [
@@ -91,7 +92,7 @@ def test_utxolocks_custom_input(server_db):
 
 
 def test_disable_utxolocks(server_db):
-    transaction.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
+    transaction_inputs.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
 
     """with `disable_utxo_locks=True` it should use the same UTXO"""
     tx1hex = construct_tx(
